Refactor filemanager stack with new api-gw construct, follow some of the SRM stack structure
