+++ config/stacks/fileManager.ts
@@ -1,4 +1,4 @@
-import { FilemanagerConfig } from '../../lib/workload/stateless/stacks/filemanager/deploy/lib/filemanager';
+import { FilemanagerConfig } from '../../lib/workload/stateless/stacks/filemanager/deploy/stack';
 import {
   AccountName,
   computeSecurityGroupName,
